@@ -16,8 +16,8 @@
         }
     ],
     "require": {
-        "php": "^7.3 || ^8.0",
-        "chrome-php/wrench": "^1.3",
+        "php": "^7.4.15 || ^8.0.2",
+        "chrome-php/wrench": "^1.5",
         "evenement/evenement": "^3.0.1",
         "monolog/monolog": "^1.27.1 || ^2.8 || ^3.2",
         "psr/log": "^1.1 || ^2.0 || ^3.0",
@@ -26,8 +26,8 @@
         "symfony/process": "^4.4 || ^5.0 || ^6.0"
     },
     "require-dev":{
-        "bamarni/composer-bin-plugin": "^1.8.1",
-        "phpunit/phpunit": "^9.5.23",
+        "bamarni/composer-bin-plugin": "^1.8.2",
+        "phpunit/phpunit": "^9.6.3 || ^10.0.12",
         "symfony/var-dumper": "^4.4 || ^5.0 || ^6.0"
     },
     "autoload":{
