@@ -18,7 +18,7 @@ use HeadlessChromium\AutoDiscover;
  */
 class AutoDiscoverTest extends BaseTestCase
 {
-    private $originalEnvPath = null;
+    private ?string $originalEnvPath = null;
 
     protected function setUp(): void
     {
@@ -48,7 +48,7 @@ class AutoDiscoverTest extends BaseTestCase
 
         $_SERVER['CHROME_PATH'] = 'test-path';
 
-        $this->assertSame($_SERVER['CHROME_PATH'], $autoDiscover->guessChromeBinaryPath());
+        self::assertSame($_SERVER['CHROME_PATH'], $autoDiscover->guessChromeBinaryPath());
     }
 
     public function testLinux(): void
@@ -57,13 +57,13 @@ class AutoDiscoverTest extends BaseTestCase
             return 'Linux';
         });
 
-        $this->assertThat(
+        self::assertThat(
             $autoDiscover->guessChromeBinaryPath(),
             $this->logicalOr(
                 'chrome',
                 'google-chrome'
             )
-          );
+        );
     }
 
     public function testMac(): void
@@ -72,7 +72,7 @@ class AutoDiscoverTest extends BaseTestCase
             return 'Darwin';
         });
 
-        $this->assertStringContainsString('.app', $autoDiscover->guessChromeBinaryPath());
+        self::assertStringContainsString('.app', $autoDiscover->guessChromeBinaryPath());
     }
 
     public function testWindows(): void
@@ -81,6 +81,6 @@ class AutoDiscoverTest extends BaseTestCase
             return 'Windows';
         });
 
-        $this->assertStringContainsString('.exe', $autoDiscover->guessChromeBinaryPath());
+        self::assertStringContainsString('.exe', $autoDiscover->guessChromeBinaryPath());
     }
 }
