@@ -11,7 +11,6 @@
 
 namespace HeadlessChromium\Browser;
 
-use HeadlessChromium\Browser;
 use HeadlessChromium\Communication\Connection;
 use HeadlessChromium\Exception\OperationTimedOut;
 use HeadlessChromium\Utils;
@@ -294,6 +293,9 @@ class BrowserProcess implements LoggerAwareInterface
             // auto debug port
             '--remote-debugging-port=0',
 
+            // allow remote access
+            '--remote-allow-origins=*',
+
             // disable undesired features
             '--disable-background-networking',
             '--disable-background-timer-throttling',
@@ -315,6 +317,11 @@ class BrowserProcess implements LoggerAwareInterface
             '--password-store=basic',
             '--use-mock-keychain', // osX only
         ];
+
+        // disable browser notifications
+        if (\array_key_exists('disableNotifications', $options) && (true === $options['disableNotifications'])) {
+            $args[] = '--disable-notifications';
+        }
 
         // enable headless mode
         if (!\array_key_exists('headless', $options) || $options['headless']) {
@@ -342,6 +349,11 @@ class BrowserProcess implements LoggerAwareInterface
             }
 
             $args[] = '--window-size='.\implode(',', $options['windowSize']);
+        }
+
+        if (\array_key_exists('userCrashDumpsDir', $options)) {
+            $args[] = '--enable-crash-reporter';
+            $args[] = '--crash-dumps-dir='.$options['userCrashDumpsDir'];
         }
 
         // sandbox mode - useful if you want to use chrome headless inside docker
@@ -432,6 +444,9 @@ class BrowserProcess implements LoggerAwareInterface
                                 $this->logger->debug('process: ✓ accepted output');
 
                                 return $matches[1];
+                            } elseif (\preg_match('/Cannot start http server for devtools\./', $output, $matches)) {
+                                $process->stop();
+                                throw new \RuntimeException('Devtools could not start');
                             } else {
                                 // log
                                 $this->logger->debug('process: ignoring output:'.\trim($output));
@@ -446,6 +461,7 @@ class BrowserProcess implements LoggerAwareInterface
 
             return Utils::tryWithTimeout($timeout, $generator($process));
         } catch (OperationTimedOut $e) {
+            $process->stop();
             throw new \RuntimeException('Cannot start browser', 0, $e);
         }
     }
