@@ -19,7 +19,6 @@ use Monolog\Handler\StreamHandler;
 use Monolog\Logger;
 use Psr\Log\LoggerInterface;
 use Psr\Log\NullLogger;
-use Symfony\Component\Process\Process;
 use Wrench\Exception\HandshakeException;
 
 class BrowserFactory
@@ -32,6 +31,7 @@ class BrowserFactory
      * - connectionDelay: Delay to apply between each operation for debugging purposes (default: none)
      * - customFlags: An array of flags to pass to the command line.
      * - debugLogger: A string (e.g "php://stdout"), or resource, or PSR-3 logger instance to print debug messages (default: none)
+     * - disableNotifications: Disable browser notifications (default: false)
      * - enableImages: Toggles loading of images (default: true)
      * - envVariables: An array of environment variables to pass to the process (example DISPLAY variable)
      * - headers: An array of custom HTTP headers
@@ -44,6 +44,7 @@ class BrowserFactory
      * - startupTimeout: Maximum time in seconds to wait for chrome to start (default: 30 sec)
      * - userAgent: User agent to use for the whole browser
      * - userDataDir: Chrome user data dir (default: a new empty dir is generated temporarily)
+     * - userCrashDumpsDir: The directory crashpad should store dumps in (crash reporter will be enabled automatically)
      * - windowSize: Size of the window. ex: `[1920, 1080]` (default: none)
      */
     protected $options = [];
