@@ -12,14 +12,16 @@ class Dom extends Node
     public function __construct(Page $page)
     {
         $message = new Message('DOM.getDocument');
-        $stream = $page->getSession()->sendMessage($message);
-        $response = $stream->waitForResponse(1000);
+        $response = $page->getSession()->sendMessageSync($message);
 
         $rootNodeId = $response->getResultData('root')['nodeId'];
 
         parent::__construct($page, $rootNodeId);
     }
 
+    /**
+     * @return Node[]
+     */
     public function search(string $selector): array
     {
         $message = new Message('DOM.performSearch', [
