@@ -28,6 +28,6 @@ class OperationTimedOut extends \Exception
             return \sprintf('%dms', (int) ($timeoutMicroSec / 1000));
         }
 
-        return \sprintf('%dμs', (int) ($timeoutMicroSec));
+        return \sprintf('%dμs', (int) $timeoutMicroSec);
     }
 }
