@@ -36,6 +36,12 @@ use HeadlessChromium\PageUtils\ResponseWaiter;
 class Page
 {
     public const DOM_CONTENT_LOADED = 'DOMContentLoaded';
+    public const FIRST_CONTENTFUL_PAINT = 'firstContentfulPaint';
+    public const FIRST_IMAGE_PAINT = 'firstImagePaint';
+    public const FIRST_MEANINGFUL_PAINT = 'firstMeaningfulPaint';
+    public const FIRST_PAINT = 'firstPaint';
+    public const INIT = 'init';
+    public const INTERACTIVE_TIME = 'InteractiveTime';
     public const LOAD = 'load';
     public const NETWORK_IDLE = 'networkIdle';
 
@@ -857,8 +863,10 @@ class Page
      *
      * @throws CommunicationException
      */
-    public function setHtml(string $html, int $timeout = 3000): void
+    public function setHtml(string $html, int $timeout = 3000, string $eventName = self::LOAD): void
     {
+        $time = \hrtime(true) / 1000 / 1000;
+
         $this->getSession()->sendMessageSync(
             new Message(
                 'Page.setDocumentContent',
@@ -866,20 +874,33 @@ class Page
                     'frameId' => $this->getFrameManager()->getMainFrame()->getFrameId(),
                     'html' => $html,
                 ]
-            )
+            ),
+            $timeout
         );
 
-        $this->waitForReload(self::LOAD, $timeout, '');
+        $timeout -= (int) \floor((\hrtime(true) / 1000 / 1000) - $time);
+
+        $this->waitForReload($eventName, \max(0, $timeout), '');
     }
 
     /**
      * Gets the raw html of the current page.
      *
      * @throws CommunicationException
+     * @throws JavascriptException
      */
     public function getHtml(?int $timeout = null): string
     {
-        return $this->evaluate('document.documentElement.outerHTML')->getReturnValue($timeout);
+        try {
+            return $this->evaluate('document.documentElement.outerHTML')->getReturnValue($timeout);
+        } catch (JavascriptException $e) {
+            if (0 === \strpos($e->getMessage(), 'Error during javascript evaluation: TypeError: Cannot read properties of null (reading \'outerHTML\')')) {
+                \usleep(1000);
+
+                return $this->evaluate('document.documentElement.outerHTML')->getReturnValue($timeout);
+            }
+            throw $e;
+        }
     }
 
     /**
