@@ -20,10 +20,7 @@ use HeadlessChromium\BrowserFactory;
  */
 class BrowsingTest extends BaseTestCase
 {
-    /**
-     * @var Browser\ProcessAwareBrowser
-     */
-    public static $browser;
+    public static Browser\ProcessAwareBrowser $browser;
 
     public static function setUpBeforeClass(): void
     {
@@ -55,12 +52,12 @@ class BrowsingTest extends BaseTestCase
         // initial navigation
         $page = $this->openSitePage('index.html');
         $title = $page->evaluate('document.title')->getReturnValue();
-        $this->assertEquals('foo', $title);
+        self::assertEquals('foo', $title);
 
         // navigate again
         $page->navigate(self::sitePath('a.html'))->waitForNavigation();
         $title = $page->evaluate('document.title')->getReturnValue();
-        $this->assertEquals('a - test', $title);
+        self::assertEquals('a - test', $title);
     }
 
     public function testFormSubmission(): void
@@ -75,7 +72,7 @@ class BrowsingTest extends BaseTestCase
         );
 
         $evaluation->waitForPageReload();
-        $this->assertEquals('hello', $page->evaluate('document.querySelector("#value").innerHTML')->getReturnValue());
+        self::assertEquals('hello', $page->evaluate('document.querySelector("#value").innerHTML')->getReturnValue());
     }
 
     public function testGetCurrentUrl(): void
@@ -84,11 +81,11 @@ class BrowsingTest extends BaseTestCase
 
         $page->getSession()->getConnection()->readData();
 
-        $this->assertEquals('about:blank', $page->getCurrentUrl());
+        self::assertEquals('about:blank', $page->getCurrentUrl());
 
         $page->navigate(self::sitePath('a.html'))->waitForNavigation();
 
-        $this->assertEquals(self::sitePath('a.html'), $page->getCurrentUrl());
+        self::assertEquals(self::sitePath('a.html'), $page->getCurrentUrl());
     }
 
     public function testPageNavigationLocalNotFoundUrl(): void
@@ -102,7 +99,7 @@ class BrowsingTest extends BaseTestCase
         // here we are testing that feature with strict and non strict modes
         $page->navigate('file:///does-not-exist')->waitForNavigation();
 
-        $this->assertTrue(true);
+        self::assertTrue(true);
     }
 
     public function testGetPages(): void
@@ -113,7 +110,7 @@ class BrowsingTest extends BaseTestCase
 
         $finalCount = \count(self::$browser->getPages());
 
-        $this->assertGreaterThan($initialCount, $finalCount);
+        self::assertGreaterThan($initialCount, $finalCount);
     }
 
     /**
@@ -130,12 +127,12 @@ class BrowsingTest extends BaseTestCase
             // initial navigation
             $page = $this->openSitePage('index.html');
             $title = $page->evaluate('document.title')->getReturnValue();
-            $this->assertEquals('foo', $title);
+            self::assertEquals('foo', $title);
 
             // navigate again
             $page->navigate(self::sitePath('a.html'))->waitForNavigation();
             $title = $page->evaluate('document.title')->getReturnValue();
-            $this->assertEquals('a - test', $title);
+            self::assertEquals('a - test', $title);
         }
     }
 
@@ -150,6 +147,6 @@ class BrowsingTest extends BaseTestCase
 
         $finalCount = \count(self::$browser->getPages());
 
-        $this->assertLessThan($initialCount, $finalCount);
+        self::assertLessThan($initialCount, $finalCount);
     }
 }
