@@ -19,10 +19,7 @@ use HeadlessChromium\BrowserFactory;
  */
 class KeyboardApiTest extends BaseTestCase
 {
-    /**
-     * @var Browser\ProcessAwareBrowser
-     */
-    public static $browser;
+    public static Browser\ProcessAwareBrowser $browser;
 
     public static function setUpBeforeClass(): void
     {
@@ -82,7 +79,7 @@ class KeyboardApiTest extends BaseTestCase
             ->evaluate('document.activeElement === document.querySelector("#myinput");')
             ->getReturnValue();
 
-        $this->assertFalse($value);
+        self::assertFalse($value);
 
         // press the Tab key
         $page->keyboard()->typeRawKey('Tab');
@@ -92,7 +89,7 @@ class KeyboardApiTest extends BaseTestCase
             ->evaluate('document.activeElement === document.querySelector("#myinput");')
             ->getReturnValue();
 
-        $this->assertTrue($value);
+        self::assertTrue($value);
     }
 
     /**
@@ -130,7 +127,7 @@ class KeyboardApiTest extends BaseTestCase
             ->getReturnValue();
 
         // check if the input contains the typed text twice
-        $this->assertEquals($text.$text, $value);
+        self::assertEquals($text.$text, $value);
     }
 
     /**
@@ -147,7 +144,7 @@ class KeyboardApiTest extends BaseTestCase
             ->press('b')
             ->release();
 
-        $this->assertEquals(0, \count($page->keyboard()->getPressedKeys()));
+        self::assertEquals(0, \count($page->keyboard()->getPressedKeys()));
     }
 
     /**
@@ -169,7 +166,7 @@ class KeyboardApiTest extends BaseTestCase
         $millisecondsElapsed = \round(\hrtime(true) / 1000 / 1000) - $start;
 
         // if this test takes less than 300ms to run (3 keys x 100ms), setKeyInterval is not working
-        $this->assertGreaterThan(300, $millisecondsElapsed);
+        self::assertGreaterThan(300, $millisecondsElapsed);
     }
 
     /**
@@ -193,6 +190,6 @@ class KeyboardApiTest extends BaseTestCase
             ->getReturnValue();
 
         // checks if the input contains the typed text
-        $this->assertSame($text, $value);
+        self::assertSame($text, $value);
     }
 }
